LayoutTests/http/tests/history/redirect-js-form-submit-before-load.html

<html>
<head>
<title>JavaScript Redirect</title>

<script>
if (window.layoutTestController) {
    layoutTestController.keepWebHistory();
    layoutTestController.waitUntilDone();
}
</script>
</head>

<body>
<p>This page is a JavaScript redirect.</p>
<form id="form" action="resources/redirect-target.html#1"></form>
<script>
document.getElementById("form").submit();
</script>
</body>
</html>
