LayoutTests/http/tests/history/redirect-js-form-submit-before-load-expected.txt

This page is the target of a redirect.

FAIL: History item count should be 1 but instead is 2.


============== Back Forward List ==============
        http://127.0.0.1:8000/history/redirect-js-form-submit-before-load.html  **nav target**
curr->  http://127.0.0.1:8000/history/resources/redirect-target.html?#1  **nav target**
===============================================
